==> inc_mort/R code/06a-attributable_cases_undernut.R <==
#' ---
#' title: PAF and attributable cases to undernutrition (BMI<18.5)
#'        As per the new systematic review of RR for undernutrition in 2024
#' author: Mathieu Bastard
#' date: 01/07/2024
#' 
#' 
#' 
#' 

rm(list=ls())

suppressPackageStartupMessages(library(data.table))
suppressPackageStartupMessages(library(propagate))
library(here)

#Run Functions script
source(here('inc_mort/R code/fun.R'))

#Function to import db from GTB databases
source(here("import/load_gtb.R"))


m <- 1e5
yr <- 2023

yrsplit <- 2020

tb <- load_gtb("tb",convert_dots = FALSE)
cty <- load_gtb("cty",convert_dots = FALSE)
pop <- load_gtb("pop",convert_dots = FALSE)
grpmbr <- load_gtb("grpmbr",convert_dots = FALSE)
sdg <- load_gtb("sdg",convert_dots = FALSE)


#load last year estiamtes data
load(here('inc_mort/estimates2023/old.rda'))
old=copy(est)
rm(est)

#load data generated by previous scripts
load(here('inc_mort/analysis/est.rda'))

#load UNAIDS data
load(here('inc_mort/analysis/unaids.rda'))

vglohi <- Vectorize(glohi, c('ev', 'sd'))


# Prevalence of BMI<18.5 in adults
# https://www.who.int/data/gho/data/indicators/indicator-details/GHO/prevalence-of-underweight-among-adults-bmi-18-(crude-estimate)-(-)
# Downloaded from the WHO Global Health Observatory using script at ~/import/save_undernutrition.R

load(here("data/gho/prev_malnut_adult.rda"))
setDT(malnutad)
setkey(malnutad, iso3, year)

malnutad2 <-
  merge(malnutad, pop[year==yr-1, .(iso3,
                              pop = e.pop.num,
                              adults.m = e.pop.m15plus,
                              adults.f = e.pop.f15plus,
                              adults = e.pop.15plus)], by = 'iso3')


# Prevalence of Zscore<-2 in children and ado
# https://www.who.int/data/gho/data/indicators/indicator-details/GHO/prevalence-of-thinness-among-children-and-adolescents-bmi--2-standard-deviations-below-the-median-(crude-estimate)-(-)
# Downloaded from the WHO Global Health Observatory using script at ~/import/save_undernutrition.R

load(here("data/gho/prev_malnut_childado.rda"))
setDT(malnutchild)
setkey(malnutchild, iso3, year)



malnutchild2 <-
  merge(malnutchild, pop[year==yr-1, .(iso3,
                              pop = e.pop.num,
                              child519=e.pop.m514+e.pop.f514+e.pop.m1524/2+e.pop.f1524/2)], by = 'iso3')


# Prevalence of Zscore<-2 in under 5
# https://www.who.int/data/gho/data/indicators/indicator-details/GHO/gho-jme-country-children-aged-5-years-wasted-br-(-weight-for-height--2-sd)
# Downloaded from the WHO Global Health Observatory using script at ~/import/save_undernutrition.R

load(here("data/gho/prev_malnut_u5.rda"))
malnutu5=setDT(malnutu5)[order(-year),head(.SD, 1) , by = iso3]
setkey(malnutu5, iso3, year)


#Replace missing lo and hi value by 25% of val
malnutu5[is.na(lo), lo:=val-1.96*0.25*val]
malnutu5[is.na(hi), hi:=val+1.96*0.25*val]

### Pull all age data together

malnutall=merge(malnutad2,malnutchild2,by="iso3")
malnutall=merge(malnutall,cty[,.(iso3,g.whoregion)],by="iso3")
malnutall=merge(malnutall,malnutu5,by="iso3")

malnutall <-
  merge(malnutall, pop[year==yr-1, .(iso3,
                                       u5 = e.pop.m04+e.pop.f04)], by = 'iso3')



### Weighted average prevalence, weighted by pop

malnutall=malnutall[,.(iso3, year=year.x, g.whoregion,prevad=val.x,prevchild=val.y,prevu5=val, prevunder=(val.x*adults+val.y*child519+val*u5)/(adults+child519+u5),
                       prevunder.lo=(lo.x*adults+lo.y*child519+lo*u5)/(adults+child519+u5),
                       prevunder.hi=(hi.x*adults+hi.y*child519+hi*u5)/(adults+child519+u5))]


rr.und <- c(1.95, (2.20-1.72) / 3.92) # undernutrition, MA review in 2024



# attributable fraction, propagating errors
#
paf2 <- function(pe, rr) {
  # @param pe proportion exposed: c(mean, sd)
  # @param rr relative risk: c(mean, sd)
  # @export
  require(propagate)
  
  DT <- cbind(pe, rr)
  EXPR <- expression(pe * (rr - 1) / (pe * (rr - 1) + 1))
  out <-
    propagate(
      expr = EXPR,
      data = DT,
      type = 'stat',
      do.sim = F,
      second.order = T
    )
  return(out)
}


out.und <- malnutall[, {
  tmp = paf2(c(prevunder/100, ((prevunder.hi/100)-(prevunder.lo/100))/3.92), rr.und)$prop
  list(paf.und = tmp[2],
       paf.und.sd = tmp[4])
}, by = .(iso3)]


malnutall=merge(malnutall,out.und,by="iso3")


malnutall2=merge(malnutall,est[year==yr-1,.(iso3,inc,inc.sd)],by="iso3")

inc.und <-
  malnutall2[, prodXY(inc, paf.und, inc.sd, paf.und.sd), by = iso3]
setnames(inc.und, c('iso3', 'inc.at.und', 'inc.at.und.sd'))


malnutall2=merge(malnutall2,inc.und,by="iso3")
malnutall2=merge(malnutall2,pop[year==yr-1,.(iso3,pop)],by="iso3")

malnutall2[, inc.at.und.num := inc.at.und * pop / m]

#Bounds
out <- malnutall2[, vglohi(inc.at.und.num, inc.at.und.sd * pop / m)]
malnutall2[, inc.at.und.lo.num := out[1,]]
malnutall2[, inc.at.und.hi.num := out[2,]]


malnutall2[,sums(inc.at.und.num)]


malnutpaf=malnutall2[,.(iso3,prev.und.new=prevunder, prev.und.sd.new=(prevunder.hi-prevunder.lo)/3.92 ,paf.und.new=paf.und,paf.und.sd.new=paf.und.sd,inc.at.und.num.new=inc.at.und.num,inc.at.und.lo.num.new=inc.at.und.lo.num,inc.at.und.hi.num.new=inc.at.und.hi.num)]


# save
#
attr(malnutpaf, "timestamp") <- Sys.Date() #set date
save(malnutpaf, file = here('inc_mort/analysis/undernutrition_new_att.rda'))
fwrite(malnutpaf, file = here(paste0('inc_mort/analysis/csv/undernutrition_new_att.rda_', Sys.Date(), '.csv')))
