==> inc_mort/R code/06b-attributable_cases_rf.R <==
#' ---
#' title: PAF and derived attributable cases to 5 risk factors by country
#' author: Mathieu Bastard
#' date: 26/06/2024


#' # Preamble
#' (Last updated: `r Sys.Date()`)
#'
#' Attributable cases by country based on the PAF approach and using the following 
#' data sources:
#' - SDG database (population exposure to selected risk factors)
#' - HIV prevalence from UNAIDS
#' - Risk ratios from litterature reviews (see below risk ratios section for
#'   details and references). The risk ratio for HIV is derived from UNAIDS estimates
#'   and TBHIV estimates
#'   
#' This file should be reran after updating age splits (Pete's folders in 
#' ~/disaggregation), which should be done if any incidence or
#' mortality estimates have been updated. (run the sequence of TB estimates, run 
#' age split scripts, rerun this script and those following in the sequence, import
#' outputs of export.R into the global TB database)
#' 
#'   
#' Deps: libraries data.table, here and propagate (propagation of errors using the delta 
#' method)
#' 
#' 
#' Input: 
#'     - est.rda (TB estimates), sdg.rda and pop.rda 
#'     - incidence splits by age from Pete Dodd datasets in ~/disaggregation/output/data
#' 
#' 
#' Output: att.rda and attributable_TIMESTAMP.csv, the related timestamped csv file
#' 
#' 
#' 
#' Note: No more than 2 sig figs in estimates of attributable cases are 
#' recommended for publishing. RRs are assumed similar for males and females, an unlikely 
#' assumption that contributes to some inconsistencies between estimates of attributable 
#' incidence by sex and totals.
#' 
#' 
#' 
#' Update 2024:
#'  - under nutrition prevalence was updated to not use anymore undernourishment prevalence and PAF+attributable cases
#'    are estimated in previous script
#'  - Keep the previous estimate using undernourishment
#'  - Update of the RR of undernutrition as per 2024 systematic review
#'  

#

rm(list=ls())

suppressPackageStartupMessages(library(data.table))
suppressPackageStartupMessages(library(propagate))
library(here)

#Run Functions script
source(here('inc_mort/R code/fun.R'))

#Function to import db from GTB databases
source(here("import/load_gtb.R"))


m <- 1e5
yr <- 2023

yrsplit <- 2020

tb <- load_gtb("tb",convert_dots = FALSE)
cty <- load_gtb("cty",convert_dots = FALSE)
pop <- load_gtb("pop",convert_dots = FALSE)
grpmbr <- load_gtb("grpmbr",convert_dots = FALSE)
sdg <- load_gtb("sdg",convert_dots = FALSE)



#load last year estiamtes data
load(here('inc_mort/estimates2023/old.rda'))
old=copy(est)
rm(est)

#load data generated by previous scripts
load(here('inc_mort/analysis/est.rda'))


#load UNDAIS data
load(here('inc_mort/analysis/unaids.rda'))



# bug fix, 07/01/2023 - something weird in the following csv file (looks like taken from last year) 
# check timestamps and use most recent
#
# incs <- fread(here('disaggregation/output/incsplits/data/A_country_incidence_disaggregated_age_sex_rate_2021-07-15.csv'))
#oincs <- fread(here('disaggregation/output/incsplits/data/A_country_incidence_disaggregated_age_sex_rate_2021-07-15.csv'))
#oincs[, V1 := NULL]


load(here('disaggregation/output/A_country_incidence_disaggregated_age_sex_num.Rdata'))
incs <- A_country_incidence_disaggregated_age_sex_num
incs <- incs[, c(3, 5:40) := lapply(.SD, function(x)as.integer(gsub(' ','', x))), .SDcols = c(3, 5:40)]
#incs <- incs[, c(16, 24:32) := lapply(.SD, function(x)as.integer(gsub(' ','', x))), .SDcols = c(16, 24:32)]
setkey(incs, iso3)
#save(incs, file = 'data/incs.Rdata')

# age <- fread(here('disaggregation/output/incsplits/data/db_estimates_country_2021-07-15.csv'))
# setkey(age, iso3)
# age[, analysed := NULL]
# save(age, file = 'Rdata/age.rdata')


# returns the last non missing value of a vector or NA
#
lastv <-
  function(x)
    return(ifelse(any(!is.na(x)), last(x[!is.na(x)]), last(x)))

# attributable fraction, propagating errors
#
paf2 <- function(pe, rr) {
  # @param pe proportion exposed: c(mean, sd)
  # @param rr relative risk: c(mean, sd)
  # @export
  require(propagate)

  DT <- cbind(pe, rr)
  EXPR <- expression(pe * (rr - 1) / (pe * (rr - 1) + 1))
  out <-
    propagate(
      expr = EXPR,
      data = DT,
      type = 'stat',
      do.sim = F,
      second.order = T
    )
  return(out)
}

# vectorized glohi
#
vglohi <- Vectorize(glohi, c('ev', 'sd'))


# # Risk Ratios (mean, sd)
#
rr.alc <-
  c(3.33, (5.19 - 2.14) / 3.92) # alcohol, Eur Resp Dis 2017
rr.dia <-
  c(1.5, (1.76 - 1.28) / 3.92) # diabetes, Trop Med Int Health 2018
rr.und <- c(1.95, (2.20-1.72) / 3.92) # undernutrition, MA review in 2024
rr.smk <- c(1.57, (2.1 - 1.18) / 3.92) # smoking, GTB 2019



# # Merge data sources
#
dta <-
  est[year == yr, .(iso3, inc, inc.lo, inc.hi, inc.sd, hiv, hiv.sd, irr, irr.sd)]

dta2 <-
  sdg[ind %in% c('alcohol',
                 'diabetes',
                 'undernutrition',
                 'smoking'), .(iso3, year, sex, ind, value, sd = (hi - lo)/3.92)]


# deps for this commented section are not met (filter function?)
#
# dta3 <-
#   reshape(
#     dta2,
#     idvar = c("iso3", "year", "sex"),
#     timevar = "ind",
#     direction = "wide"
#   ) %>%
#   filter(year <= yr)
# names(dta3) <- gsub('value.', '', names(dta3))
# setkey(dta3, iso3)

dta3 <-
  reshape(
    dta2[year<=yr],
    idvar = c("iso3", "year", "sex"),
    timevar = "ind",
    direction = "wide"
  ) 
names(dta3) <- gsub('value.', '', names(dta3))
setkey(dta3, iso3, sex)



dta4 <- dta3[, .(
  diabetes = lastv(diabetes) / 100,
  alcohol = lastv(alcohol) / 100,
  undernutrition = lastv(undernutrition) / 100,
  smoking = lastv(smoking) / 100,
  diabetes.sd = lastv(sd.diabetes) / 100,
  alcohol.sd = lastv(sd.alcohol) / 100,
  undernutrition.sd = lastv(sd.undernutrition) / 100,
  smoking.sd = lastv(sd.smoking) / 100
),
by = .(iso3, sex)]

incs2 <-
  merge(incs, pop[year==yr, .(iso3,
                              pop = e.pop.num,
                              pop.m = e.pop.m04 + e.pop.m514 + e.pop.m15plus,
                              pop.f = e.pop.f04 + e.pop.f514 + e.pop.f15plus,
                              adults.m = e.pop.m15plus,
                              adults.f = e.pop.f15plus,
                              adults = e.pop.15plus)], by = 'iso3')
incs3 <- incs2[, .(iso3,
                  inc.m = inc.num.m/pop * m,
                  inc.m.sd = ((inc.num.m.hi-inc.num.m.lo)/3.92)/pop * m,
                  inc.f = inc.num.f/pop.f * m,
                  inc.f.sd = ((inc.num.f.hi-inc.num.f.lo)/3.92)/pop.f * m,
                  inc.15plus = inc.num.15plus/adults * m,
                  inc.15plus.sd = ((inc.num.15plus.hi-inc.num.15plus.lo)/3.92)/adults * m,
                  inc.15plus.m = inc.num.15plus.m/adults.m * m,
                  inc.15plus.m.sd = ((inc.num.15plus.m.hi-inc.num.15plus.m.lo)/3.92)/adults.m * m,
                  inc.15plus.f = inc.num.15plus.f/adults.f * m,
                  inc.15plus.f.sd = ((inc.num.15plus.f.hi-inc.num.15plus.f.lo)/3.92)/adults.f * m,
                  pop,pop.m,pop.f,adults,adults.m,adults.f)]

dta5 <-
  merge(dta4, incs3, by = 'iso3')
dim(dta4)
dim(dta5)


# inelegant reshaping
dta5[sex=='m', pop := pop.m]
dta5[sex=='f', pop := pop.f]
dta5[, pop.m := NULL]
dta5[, pop.f := NULL]

dta5[sex=='m', adults := adults.m]
dta5[sex=='f', adults := adults.f]
dta5[, adults.m := NULL]
dta5[, adults.f := NULL]

# dta5[sex=='m', adults18 := adults18.m]
# dta5[sex=='f', adults18 := adults18.f]
# dta5[, adults18.m := NULL]
# dta5[, adults18.f := NULL]

dta5[sex=='m', inc.15plus := inc.15plus.m]
dta5[sex=='f', inc.15plus := inc.15plus.f]
dta5[, inc.15plus.m := NULL]
dta5[, inc.15plus.f := NULL]

dta5[sex=='m', inc.15plus.sd := inc.15plus.m.sd]
dta5[sex=='f', inc.15plus.sd := inc.15plus.f.sd]
dta5[, inc.15plus.m.sd := NULL]
dta5[, inc.15plus.f.sd := NULL]


ac <- merge(dta5, dta, by = 'iso3', all.x = TRUE)

ac[sex=='m', inc := inc.m]
ac[sex=='f', inc := inc.f]
ac[, inc.m := NULL]
ac[, inc.f := NULL]

ac[sex=='m', inc.sd := inc.m.sd]
ac[sex=='f', inc.sd := inc.f.sd]
ac[, inc.m.sd := NULL]
ac[, inc.f.sd := NULL]

ac[, inc.lo := NULL]
ac[, inc.hi := NULL]

ac[sex %ni% 'a', hiv := NA]
ac[sex %ni% 'a', hiv.sd := NA]
ac[sex %ni% 'a', irr := NA]
ac[sex %ni% 'a', irr.sd := NA]



# # Attributable fractions
#

# HIV
#
out.hiv <- ac[, {
  tmp = paf2(c(hiv, hiv.sd), c(irr, irr.sd))$prop
  list(paf.hiv = tmp[2],
       paf.hiv.sd = tmp[4])
}, by = .(iso3, sex)]

out.hiv[paf.hiv<0, paf.hiv := 0]


# diabetes
#
out.dia <- ac[, {
  tmp = paf2(c(diabetes, diabetes.sd), rr.dia)$prop
  list(paf.dia = tmp[2],
       paf.dia.sd = tmp[4])
}, by = .(iso3, sex)]


# alcohol
#
out.alc <- ac[, {
  tmp = paf2(c(alcohol, alcohol.sd), rr.alc)$prop
  list(paf.alc = tmp[2],
       paf.alc.sd = tmp[4])
}, by = .(iso3, sex)]

out.alc[paf.alc<0, paf.alc := 0]


# smoking
#
out.smk <- ac[, {
  tmp = paf2(c(smoking, smoking.sd), rr.smk)$prop
  list(paf.smk = tmp[2],
       paf.smk.sd = tmp[4])
}, by = .(iso3, sex)]

out.smk[paf.smk<0, paf.smk := 0]


# undernutrition
#
out.und <- ac[, {
  tmp = paf2(c(undernutrition, undernutrition.sd), rr.und)$prop
  list(paf.und = tmp[2],
       paf.und.sd = tmp[4])
}, by = .(iso3, sex)]


ac <-
  cbind(
    ac,
    out.hiv[, -1],
    out.dia[, -1],
    out.alc[, -1],
    out.smk[, -1],
    out.und[, -1]
  )




# # Attributable incidence rate
#
inc.hiv <-
  ac[, prodXY(inc, paf.hiv, inc.sd, paf.hiv.sd), by = iso3]
setnames(inc.hiv, c('iso3', 'inc.at.hiv', 'inc.at.hiv.sd'))
# inc.hiv[, inc.at.hiv.sd := inc.at.hiv.sd]

inc.dia <-
  ac[, prodXY(inc.15plus, paf.dia, inc.15plus.sd, paf.dia.sd), by = iso3]
setnames(inc.dia, c('iso3', 'inc.at.dia', 'inc.at.dia.sd'))
# inc.dia[, inc.at.dia.sd := inc.at.dia.sd]

inc.alc <-
  ac[, prodXY(inc.15plus, paf.alc, inc.15plus.sd, paf.alc.sd), by = iso3]
setnames(inc.alc, c('iso3', 'inc.at.alc', 'inc.at.alc.sd'))
# inc.alc[, inc.at.alc.sd := inc.at.alc.sd]

inc.smk <-
  ac[, prodXY(inc.15plus, paf.smk, inc.15plus.sd, paf.smk.sd), by = iso3]
setnames(inc.smk, c('iso3', 'inc.at.smk', 'inc.at.smk.sd'))
# inc.smk[, inc.at.smk.sd := inc.at.smk.sd]

inc.und <-
  ac[, prodXY(inc, paf.und, inc.sd, paf.und.sd), by = iso3]
setnames(inc.und, c('iso3', 'inc.at.und', 'inc.at.und.sd'))
# inc.und[, inc.at.und.sd := inc.at.und.sd]

ac <-
  cbind(
    ac,
    inc.hiv[, -1],
    inc.dia[, -1],
    inc.alc[, -1],
    inc.smk[, -1],
    inc.und[, -1]
  )




# # Attributable cases
#

# drop at.hiv values where tbhiv < 1e-4 (inc.h very close to zero)
# (inconsistencies due to estimated mean of ratios of SDs being 
# too different from ratios of means of SDs)
lst <- est[year==yr & tbhiv<1e-4, iso3]
ac[iso3 %in% lst, inc.at.hiv := NA]
ac[iso3 %in% lst, inc.at.hiv.sd := NA]

ac[, inc.at.hiv.num := inc.at.hiv * pop / m]
ac[, inc.at.dia.num := inc.at.dia * adults / m]
ac[, inc.at.alc.num := inc.at.alc * adults / m]
ac[, inc.at.smk.num := inc.at.smk * adults / m]
ac[, inc.at.und.num := inc.at.und * pop / m]




# bounds
#
sel <- !is.na(ac$inc.at.hiv)
out <- ac[sel, vglohi(inc.at.hiv.num, inc.at.hiv.sd * pop / m)]
ac[sel, inc.at.hiv.lo.num := out[1,]]
ac[sel, inc.at.hiv.hi.num := out[2,]]


sel <- !is.na(ac$inc.at.dia) & ac$inc.at.dia.sd > 0
out <- ac[sel, vglohi(inc.at.dia.num, inc.at.dia.sd * pop / m)]
ac[sel, inc.at.dia.lo.num := out[1,]]
ac[sel, inc.at.dia.hi.num := out[2,]]


sel <- !is.na(ac$inc.at.alc) & ac$inc.at.alc.sd > 0
out <- ac[sel, vglohi(inc.at.alc.num, inc.at.alc.sd * pop / m)]
ac[sel, inc.at.alc.lo.num := out[1,]]
ac[sel, inc.at.alc.hi.num := out[2,]]


sel <- !is.na(ac$inc.at.smk) & ac$inc.at.smk.sd > 0
out <- ac[sel, vglohi(inc.at.smk.num, inc.at.smk.sd * pop / m)]
ac[sel, inc.at.smk.lo.num := out[1,]]
ac[sel, inc.at.smk.hi.num := out[2,]]


sel <- !is.na(ac$inc.at.und) & ac$inc.at.und.sd > 0
out <- ac[sel, vglohi(inc.at.und.num, inc.at.und.sd * pop / m)]
ac[sel, inc.at.und.lo.num := out[1,]]
ac[sel, inc.at.und.hi.num := out[2,]]

ac[, 46:60 := lapply(.SD, function(x)
  as.integer(x)), .SDcols = 46:60]

ac[inc.at.smk.hi.num < inc.at.smk.num, inc.at.smk.hi.num := as.integer(3 * inc.at.smk.num)]




# checks
#
ac[inc.at.dia.num>0, test.bounds(inc.at.dia.num, inc.at.dia.lo.num, inc.at.dia.hi.num)]
ac[inc.at.alc.num>0, test.bounds(inc.at.alc.num, inc.at.alc.lo.num, inc.at.alc.hi.num)]
ac[inc.at.smk.num>0, test.bounds(inc.at.smk.num, inc.at.smk.lo.num, inc.at.smk.hi.num)]
ac[inc.at.hiv.num>0, test.bounds(inc.at.hiv.num, inc.at.hiv.lo.num, inc.at.hiv.hi.num)]
ac[inc.at.und.num>0, test.bounds(inc.at.und.num, inc.at.und.lo.num, inc.at.und.hi.num)]

att <- copy(ac)


### Merge new attributable cases based on under nutrition and not undernourishment


load(here("inc_mort/analysis/undernutrition_new_att.rda"))
malnutpaf[,sex:="a"]
att$sex.1=NULL

#duplicates of sex variables to remove
att$sex=NULL
att$sex=NULL
att$sex=NULL
att$sex=NULL
att$sex=NULL

att2=merge(att,malnutpaf,by=c("iso3","sex"))


att <- copy(att2)


# check aggregates
#
att[sex=='a', sums(inc.at.und.num)]
att[sex=='a', sums(inc.at.und.num.new)]

att[sex=='a', sums(inc.at.dia.num)]
att[sex=='a', sums(inc.at.smk.num)]
att[sex=='a', sums(inc.at.hiv.num)]
att[sex=='a', sums(inc.at.alc.num)]


rm(dta, dta2, dta3, dta4, dta5)


att$year=yr

# save
#
attr(att, "timestamp") <- Sys.Date() #set date
save(att, file = here('inc_mort/analysis/attributable_cases.rda'))
fwrite(att, file = here(paste0('inc_mort/analysis/csv/attributable_cases_', Sys.Date(), '.csv')))






# # Attributable cases, RF tables
#
att2 <- att[, list(
  iso3,
  sex,
  inc.at.hiv.num,
  inc.at.hiv.lo.num,
  inc.at.hiv.hi.num,
  inc.at.dia.num,
  inc.at.dia.lo.num,
  inc.at.dia.hi.num,
  inc.at.alc.num,
  inc.at.alc.lo.num,
  inc.at.alc.hi.num,
  inc.at.smk.num,
  inc.at.smk.lo.num,
  inc.at.smk.hi.num,
  inc.at.und.num,
  inc.at.und.lo.num,
  inc.at.und.hi.num,
  inc.at.und.num.new,
  inc.at.und.lo.num.new,
  inc.at.und.hi.num.new,
  sex
)]


# reshape to long
#
att3 <- melt(att2, id.vars = c(1,2, 4:5), measure.vars = 3)
att4 <- melt(att2, id.vars = c(1,2, 7:8), measure.vars = 6)
att5 <- melt(att2, id.vars = c(1,2, 10:11), measure.vars = 9)
att6 <- melt(att2, id.vars = c(1,2, 13:14), measure.vars = 12)
att7 <- melt(att2, id.vars = c(1,2, 16:17), measure.vars = 15)

att8 <- melt(att2, id.vars = c(1,2, 19:20), measure.vars = 18)

for (i in list(att3, att4, att5, att6, att7,att8)) setnames(i, c('iso3', 'sex', 'lo', 'hi', 'risk.factor', 'best'))
att3[, risk.factor := 'hiv']
att4[, risk.factor := 'diabetes']
att5[, risk.factor := 'alcohol']
att6[, risk.factor := 'smoking']
att7[, risk.factor := 'undernutrition old']
att8[, risk.factor := 'undernutrition']

rf <-
  rbind(att3, att4, att5, att6, att7,att8)
rf[, measure := 'inc']
rf[, unit := 'num']
setkey(rf, iso3)

rf[, age := 'a']
rf[risk.factor %in% c('smoking', 'alcohol'), age := '15+']
rf[risk.factor %in% c('diabetes'), age := '18+']

rf <- merge(rf, cty[, .(iso3, g.whoregion)], by = 'iso3')

sel <- rf$risk.factor %in% c('hiv', 'undernutrition old',"undernutrition") & rf$sex %in% c('m','f')
table(sel)
rf <- rf[!sel]


#Save

rf$year=yr
attr(rf, "timestamp") <- Sys.Date() #set date
save(rf, file = here('inc_mort/analysis/rf.rda'))

fwrite(rf[, .(iso3, year = yr, risk.factor, measure, unit, sex, age, best, lo, hi)],
       file = here(paste0('inc_mort/analysis/csv/db_inc_risk_factor_country_', Sys.Date(), '.csv')))



# aggregate rf
#
vglohi <- Vectorize(glohi, c('ev', 'sd'))

rf.regional <- rf[, .(best = as.integer(sums(best)),
                      sd = as.integer(sqrt(sums( ((hi - lo) / 3.92) ^ 2) ))),
                  by = .(g.whoregion, risk.factor, measure, unit, sex, age)]

out <- with(rf.regional, vglohi(best, sd))
rf.regional[, lo := as.integer(out[1,])]
rf.regional[, hi := as.integer(out[2,])]
rf.regional[, sd := NULL]

rf.global <- rf[, .(best = as.integer(sums(best)),
                    sd = as.integer(sqrt(sums( ((hi - lo) / 3.92) ^ 2) ))),
                by = .(risk.factor, measure, unit, sex, age)]

out <- with(rf.global, vglohi(best, sd))
rf.global[, lo := as.integer(out[1,])]
rf.global[, hi := as.integer(out[2,])]
rf.global[, sd := NULL]

rf.regional[, group_type := 'g_whoregion']
rf.regional[, group_name := g.whoregion]
rf.regional[, g.whoregion := NULL]
rf.regional[, year := yr]

rf.global[, group_type := 'global']
rf.global[, group_name := 'global']
rf.global[, year := yr]

cols <- c(
  'group_type',
  'group_name',
  'year',
  'risk.factor',
  'measure',
  'unit',
  'sex',
  'age',
  'best',
  'lo',
  'hi'
)

setcolorder(rf.global, cols)
setcolorder(rf.regional, cols)


rf.global$year=yr
rf.regional$year=yr

attr(rf.global, "timestamp") <- Sys.Date() 
attr(rf.regional, "timestamp") <- Sys.Date()

save(rf.global, file = here('inc_mort/analysis/rf.global.rda'))
save(rf.regional, file = here('inc_mort/analysis/rf.regional.rda'))
