==> inc_mort/R code/07-incidence_bact_confirmed.R <==
#' ---
#' title: Incidence of bacteriologically confirmed pulmonary TB
#' author: Mathieu Bastard
#' date: 03/07/2024

#' This script estimates the incidence of TB that would be found bacteriologically
#' confirmed if all incident cases were tested using recommended modern diagnostics.
#' The estimate is calculated as the product of incidence with the average 
#' proportion bacteriologically-confirmed among notified new cases in the 
#' subset of high-income countries. 
#' 
#' deps: libraries data.table and here as for other scripts in the sequence
#' 
#' 
#' input data:
#' - est.rda (estimates) 
#' - tb.rda (notification data for its data on Bact confirmation)
#' 
#' 
#' output: bpos.rda and related timestamped csv data
#'
#'
#' Note: the output from this file was not used in Global TB report 2022
#' 


rm(list=ls())

library(data.table)
library(here)

#Run Functions script
source(here('inc_mort/R code/fun.R'))

#Function to import db from GTB databases
source(here("import/load_gtb.R"))


m <- 1e5
yr <- 2023

yrsplit <- 2020

tb <- load_gtb("tb",convert_dots = FALSE)
cty <- load_gtb("cty",convert_dots = FALSE)
pop <- load_gtb("pop",convert_dots = FALSE)
grpmbr <- load_gtb("grpmbr",convert_dots = FALSE)
sdg <- load_gtb("sdg",convert_dots = FALSE)


#load data generated by previous scripts
load(here('inc_mort/analysis/est.rda'))

# vectorized lohi
#
vlohi <- Vectorize(lohi, c('ev', 'sd'))


# weighted average b+ by income group and year
#
tb <- merge(tb, est[year == yr, .(iso3, g.income)], by = 'iso3')
tb[year > 2013, conf.denom := rowSums(cbind(new.ep, ret.rel.ep), na.rm = TRUE)]
out <-
  tb[year > 2017 &
       !is.na(g.income) &
       c.newinc > 1000, .(conf = weighted.mean(conf, w = conf.denom, na.rm = T)), by =
       .(g.income, year)]
out.sd <-
  tb[year > 2017 &
       !is.na(g.income) &
       c.newinc > 1000, .(conf.sd = weighted.sd(conf, w = conf.denom, na.rm = T)), by =
       .(g.income, year)]

tb[year > 2017 &
     !is.na(g.income) &
     c.newinc > 1000, .(min(conf)), by = .(g.income, year)]
tb[year > 2017 &
     !is.na(g.income) &
     c.newinc > 1000, .(max(conf)), by = .(g.income, year)]


# use HIC values of conf to estimate incidence B+
#
hic <- last(out$conf)
hic.sd <- last(out.sd$conf.sd)

est2 <- merge(est[year==yr, .(iso3,pop,g.whoregion,g.income,inc,inc.sd)], tb[year==yr, .(iso3,ep)], by='iso3')

# impute missing ep
#
out <- est2[, .(ep.imp = mean(ep, na.rm=T)), by=g.whoregion]
est3 <- merge(est2, out, by='g.whoregion')
est3[is.na(ep), ep := ep.imp]

bpos <- est3[, {
  tmp = prodXY(inc * (1 - ep), hic, inc.sd * (1 - ep), hic.sd)

  list (inc.bc = tmp[[1]],
        inc.bc.sd = tmp[[2]])
},
by = .(iso3)]

sel <- bpos$inc.bc > 0
out <- vlohi(bpos$inc.bc[sel] / m, bpos$inc.bc.sd[sel] / m)

bpos$inc.bc.lo[sel] <- out[1, ] * m
bpos$inc.bc.hi[sel] <- out[2, ] * m

bpos[!sel]
bpos[sel, test.bounds(inc.bc, inc.bc.lo, inc.bc.hi)]
bpos <- merge(bpos, est[year == yr, .(iso3, inc)], by = 'iso3')
bpos[, test.AgeB(inc, inc.bc)]
bpos[, inc := NULL]


attr(bpos, "timestamp") <- Sys.Date() 
save(bpos, file = here('inc_mort/analysis/inc_bpos.rda'))
fwrite(bpos, file = here(paste0('inc_mort/analysis/csv/inc_bpos_', Sys.Date(), '.csv')))
